requirements.txt
PyQt6>=6.6.0
PyQt6-Charts>=6.6.0
mysql-connector-python>=8.2.0
bcrypt>=4.1.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
matplotlib>=3.8.0
pandas>=2.1.0
reportlab>=4.0.7
Pillow>=10.1.0
